Rename and create Navigation class
